Adds deprecation grace period to diffing
Extends schema contract diffing to include a grace period for deprecations.

This allows for a temporary grace period before a deprecated feature is considered a breaking change, providing more flexibility in managing API evolution.
A new `DEPRECATION_GRACE` ChangeType and corresponding ClassificationCount property are added to represent this grace period.
Moves ClassificationCount interface to a shared location for use by both schema-contract and tools/schema modules.
